Refactor: rename arrow_table.py + friends (#3459)
- Renames `streamlit/elements/arrow_table.py` -> `streamlit/components/v1/component_arrow.py`
- Consolidates all Arrow-related protobuf definitions inside `Components.proto` (alongside all the other Components bits).

All of these bits are specific to Components dataframe serialization. The upcoming "use Arrow for all dataframe serialization" feature _also_ uses Arrow, of course - but it doesn't share any of this logic. So this is mostly a "reduce incoming confusion" refactor.
