+++ lib/streamlit/components/v1/component_arrow.py
@@ -12,6 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+"""Data marshalling utilities for ArrowTable protobufs, which are used by
+CustomComponent for dataframe serialization.
+"""
+
 import pandas as pd
 import pyarrow as pa
 from streamlit import util
