+++ proto/streamlit/proto/Element.proto
@@ -39,7 +39,7 @@ import "streamlit/proto/NumberInput.proto";
 import "streamlit/proto/Markdown.proto";
 import "streamlit/proto/MultiSelect.proto";
 import "streamlit/proto/PlotlyChart.proto";
-import "streamlit/proto/ComponentInstance.proto";
+import "streamlit/proto/Components.proto";
 import "streamlit/proto/Progress.proto";
 import "streamlit/proto/Radio.proto";
 import "streamlit/proto/Selectbox.proto";
